fix: move injection tokens to config-provider
