+++ packages/frameworks/angular/projects/renderer/src/lib/config-provider/index.ts
@@ -1,1 +1,2 @@
 export * from './config-provider.component';
+export * from './injection-tokens';
