+++ packages/frameworks/angular/projects/renderer/src/public-api.ts
@@ -3,7 +3,6 @@
  */
 
 export * from './lib/genui-renderer';
-export * from './lib/injection-tokens';
 export * from './lib/config-provider';
 export * from './lib/legacy-renderer';
 export { RENDERER_SETTINGS } from '@opentiny/tiny-schema-renderer-ng';
